feat: Apply remarks from bots reviewers

+++ lib/features/diary/presentation/widgets/day_info_widget.dart
@@ -282,7 +282,7 @@ class DayInfoWidget extends StatelessWidget {
       builder: (context) => const DeleteDialog(),
     );
 
-    if (deleteWeight != null) {
+    if (deleteWeight != null && userWeightEntity != null) {
       onDeleteWeight(userWeightEntity!, trackedDayEntity);
     }
   }
